server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
httpx==0.27.0
json5>=0.9.0

# MongoDB
motor==3.3.2
pymongo==4.6.0
beanie==1.25.0

# Auth0 / JWT Authentication
PyJWT==2.8.0
cryptography==41.0.7
